+++ mobile/android/android-components/components/feature/addons/src/main/java/mozilla/components/feature/addons/ui/PermissionsDialogFragment.kt
@@ -135,7 +135,7 @@ class PermissionsDialogFragment : AppCompatDialogFragment() {
         rootView.findViewById<TextView>(R.id.title).text =
             requireContext().getString(
                 R.string.mozac_feature_addons_permissions_dialog_title,
-                addon.translatedName
+                addon.translateName(requireContext())
             )
         rootView.findViewById<TextView>(R.id.permissions).text = buildPermissionsText()