+++ mobile/android/android-components/samples/browser/src/main/java/org/mozilla/samples/browser/addons/AddonsFragment.kt
@@ -25,7 +25,7 @@ import mozilla.components.feature.addons.ui.AddonInstallationDialogFragment
 import mozilla.components.feature.addons.ui.AddonsManagerAdapter
 import mozilla.components.feature.addons.ui.AddonsManagerAdapterDelegate
 import mozilla.components.feature.addons.ui.PermissionsDialogFragment
-import mozilla.components.feature.addons.ui.translatedName
+import mozilla.components.feature.addons.ui.translateName
 import org.mozilla.samples.browser.R
 import org.mozilla.samples.browser.ext.components
 import java.util.concurrent.CancellationException
@@ -197,7 +197,7 @@ class AddonsFragment : Fragment(), AddonsManagerAdapterDelegate {
                     Toast.makeText(
                         requireContext(), getString(
                         R.string.mozac_feature_addons_failed_to_install,
-                        addon.translatedName
+                        addon.translateName(requireContext())
                 ),
                         Toast.LENGTH_SHORT
                     ).show()