+++ mobile/android/android-components/samples/browser/src/main/java/org/mozilla/samples/browser/addons/InstalledAddonDetailsActivity.kt
@@ -15,7 +15,7 @@ import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
 import mozilla.components.feature.addons.Addon
 import mozilla.components.feature.addons.AddonManagerException
-import mozilla.components.feature.addons.ui.translatedName
+import mozilla.components.feature.addons.ui.translateName
 import org.mozilla.samples.browser.ext.components
 import org.mozilla.samples.browser.BrowserActivity
 import org.mozilla.samples.browser.R
@@ -61,7 +61,7 @@ class InstalledAddonDetailsActivity : AppCompatActivity() {
     }
 
     private fun bindUI(addon: Addon) {
-        title = addon.translatedName
+        title = addon.translateName(this)
 
         bindEnableSwitch(addon)
 
@@ -87,14 +87,14 @@ class InstalledAddonDetailsActivity : AppCompatActivity() {
                         switch.setState(true)
                         Toast.makeText(
                             this,
-                            getString(R.string.mozac_feature_addons_successfully_enabled, addon.translatedName),
+                            getString(R.string.mozac_feature_addons_successfully_enabled, addon.translateName(this)),
                             Toast.LENGTH_SHORT
                         ).show()
                     },
                     onError = {
                         Toast.makeText(
                             this,
-                            getString(R.string.mozac_feature_addons_failed_to_enable, addon.translatedName),
+                            getString(R.string.mozac_feature_addons_failed_to_enable, addon.translateName(this)),
                             Toast.LENGTH_SHORT
                         ).show()
                     }
@@ -106,14 +106,14 @@ class InstalledAddonDetailsActivity : AppCompatActivity() {
                         switch.setState(false)
                         Toast.makeText(
                             this,
-                            getString(R.string.mozac_feature_addons_successfully_disabled, addon.translatedName),
+                            getString(R.string.mozac_feature_addons_successfully_disabled, addon.translateName(this)),
                             Toast.LENGTH_SHORT
                         ).show()
                     },
                     onError = {
                         Toast.makeText(
                             this,
-                            getString(R.string.mozac_feature_addons_failed_to_disable, addon.translatedName),
+                            getString(R.string.mozac_feature_addons_failed_to_disable, addon.translateName(this)),
                             Toast.LENGTH_SHORT
                         ).show()
                     }
@@ -177,7 +177,7 @@ class InstalledAddonDetailsActivity : AppCompatActivity() {
                 onSuccess = {
                     Toast.makeText(
                         this,
-                        getString(R.string.mozac_feature_addons_successfully_uninstalled, addon.translatedName),
+                        getString(R.string.mozac_feature_addons_successfully_uninstalled, addon.translateName(this)),
                         Toast.LENGTH_SHORT
                     ).show()
                     finish()
@@ -185,7 +185,7 @@ class InstalledAddonDetailsActivity : AppCompatActivity() {
                 onError = { _, _ ->
                     Toast.makeText(
                         this,
-                        getString(R.string.mozac_feature_addons_failed_to_uninstall, addon.translatedName),
+                        getString(R.string.mozac_feature_addons_failed_to_uninstall, addon.translateName(this)),
                         Toast.LENGTH_SHORT
                     ).show()
                 }